+++ prog/opencl_module.h
@@ -318,7 +318,7 @@ public:
 	 *
 	 * @param in Name of the kernel under consideration.
 	 */
-	virtual int get_flop_size(const char * in);
+	virtual uint64_t get_flop_size(const char * in);
 
 	/**
 	 * Return amount of bytes read and written by a specific kernel per call.
@@ -337,7 +337,7 @@ public:
 	 * @param read_write_size number of bytes read and written by the kernel
 	 * @param flop_size amount of flops performed by the kernel
 	 */
-	void print_profiling(std::string filename, const char * kernelName, uint64_t time_total, int calls_total, size_t read_write_size, int flop_size);
+	void print_profiling(std::string filename, const char * kernelName, uint64_t time_total, int calls_total, size_t read_write_size, uint64_t flop_size);
 
 	/**
 	 * Enqueue the given kernel on the device. Local work size will be determined
