+++ prog/opencl_module_fermions.cpp
@@ -1971,12 +1971,12 @@ int flop_dslash_per_site(inputparameters * parameters)
 
 }
 
-int Opencl_Module_Fermions::get_flop_size(const char * in)
+uint64_t Opencl_Module_Fermions::get_flop_size(const char * in)
 {
-	int result = Opencl_Module_Spinors::get_flop_size(in);
+	uint64_t result = Opencl_Module_Spinors::get_flop_size(in);
 	if (result != 0) return result;
-	int S = get_parameters()->get_spinorfieldsize();
-	int Seo = get_parameters()->get_eoprec_spinorfieldsize();
+	uint64_t S = get_parameters()->get_spinorfieldsize();
+	uint64_t Seo = get_parameters()->get_eoprec_spinorfieldsize();
 	if (strcmp(in, "M_wilson") == 0) {
 		//this kernel performs one dslash on each site and adds this to a spinor
 		return S * (flop_dslash_per_site(get_parameters()) + NC * NDIM * get_parameters()->get_flop_complex_mult() + NC * NDIM * 2 );
