+++ prog/opencl_module_fermions.h
@@ -288,7 +288,7 @@ public:
 	 *
 	 * @param in Name of the kernel under consideration.
 	 */
-	virtual int get_flop_size(const char * in);
+	virtual uint64_t get_flop_size(const char * in);
 
 private:
 	////////////////////////////////////
