+++ prog/opencl_module_heatbath.h
@@ -129,7 +129,7 @@ private:
 	 *
 	 * @param in Name of the kernel under consideration.
 	 */
-	virtual int get_flop_size(const char * in);
+	virtual uint64_t get_flop_size(const char * in);
 
 
 };
